Update tracer to use wrapper func over standard OpenAI Client
