+++ judgeval/common/tracer.py
@@ -323,39 +323,37 @@ class Tracer:
             
         return wrapper
 
-
-class TracedOpenAI(OpenAI):
+def wrap_openai(client: OpenAI) -> OpenAI:
     """
-    OpenAI client that automatically traces all API calls.
+    Wraps an OpenAI client to add tracing capabilities.
     Usage:
-        client = TracedOpenAI()
-        response = client.chat.completions.create(...)
+        client = OpenAI()
+        traced_client = wrap_openai(client)
     """
-    def __init__(self, *args, **kwargs):
-        super().__init__(*args, **kwargs)
-        self.tracer = Tracer._instance  # Get the global tracer instance
-        
-        # Store the original create method
-        original_create = self.chat.completions.create
-        
-        # Replace with our traced version
-        def traced_create(*args, **kwargs):
-            if self.tracer and self.tracer._current_trace:
-                with self.tracer._current_trace.span("llm_call") as span:
-                    # Record the input
-                    span.record_input({
-                        "model": kwargs.get("model"),
-                        "messages": kwargs.get("messages")
-                    })
-                    
-                    # Make the API call
-                    response = original_create(*args, **kwargs)
-                    
-                    # Record the output
-                    span.record_output(response.choices[0].message.content)
-                    return response
-            else:
-                # If no trace context, just call the original method
-                return original_create(*args, **kwargs)
+    tracer = Tracer._instance  # Get the global tracer instance
+    
+    # Store the original create method
+    original_create = client.chat.completions.create
+    
+    # Replace with our traced version
+    def traced_create(*args, **kwargs):
+        if tracer and tracer._current_trace:
+            with tracer._current_trace.span("llm_call") as span:
+                # Record the input
+                span.record_input({
+                    "model": kwargs.get("model"), 
+                    "messages": kwargs.get("messages")
+                })
                 
-        self.chat.completions.create = traced_create
+                # Make the API call
+                response = original_create(*args, **kwargs)
+                
+                # Record the output
+                span.record_output(response.choices[0].message.content)
+                return response
+        else:
+            # If no trace context, just call the original method
+            return original_create(*args, **kwargs)
+            
+    client.chat.completions.create = traced_create
+    return client
